refactor: organize CreationSteps into per-step subdirectories
Step1/ — Step1, GameSidebar, TemplateBrowser
Step2/ — Step2, TemplateHeader
Step3/ — Step3, DockerImageSection, HardwareLimitsSection

Update all relative import paths accordingly.
